integration-service/requirements.txt
flask
pymysql
mysql-connector-python